# apnashift-backend/openapi.yaml
openapi: 3.0.3
info:
  title: ApnaShift Backend (MVP)
  description: On-demand shifting API — pickup, mini truck, mini tractor. Auth via Bearer JWT.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [ok, error]
      properties:
        ok: { type: boolean, example: false }
        error: { type: string, example: invalid_credentials }
...
    PublicUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        phone: { type: string }
        role: { type: string, enum: [user, driver, admin] }
...
paths:
...
  /api/auth/register:
    post:
      summary: Register user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, phone, password]
              properties:
                name: { type: string, example: Rahul Sharma }
                phone: { type: string, example: '98765 43210' }
                password: { type: string, minLength: 8, example: password123 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
                  token: { type: string }
                  user: { $ref: '#/components/schemas/PublicUser' }
        '409': { description: Phone taken (phone_taken), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/auth/login:
    post:
      summary: Login (user/driver/admin by phone)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, password]
              properties:
                phone: { type: string, example: '9876543210' }
                password: { type: string, example: password123 }
      responses:
        '200':
          description: OK, JWT with {id, role}
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
                  token: { type: string }
                  user: { $ref: '#/components/schemas/PublicUser' }
        '401': { description: invalid_credentials (wrong phone or password, same), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...